Absolute Mouse Movements
This implements a mouse cursor that snaps to the position of the analog stick. Useful for aiming in some mouse-based 2d side scrollers or perhaps some 3d games where you'd like to have a "look" option.

Because the ioctl for absolute mouse movement can't be layered onto a relative mouse file descriptor, I broke out three uinput devices: xbox controller, regular mouse and keyboard, and absolute mouse.  Which device is used for which operation is a bit hard-coded at the moment but it all seems to work.