Revert Xbox uinput setup to deprecated method
This fixes bugs causing the uinput device to not initialize correctly, resulting in a corrupted device name, missing ABS values and the controller not being recognised correctly by SDL2. I chose to revert back to the deprecated method rather than upgrading the rest of the setup to the new method so that CFWs on old kernels remain compatible.