Fix fake xbox controller name filtering
This fixes the code that was supposed to stop gptokeyb2 from reading its own controller input. SDL_GameControllerNameForIndex returns the name as specified by gamecontrollerdb.txt (Microsoft Xbox 360), not the actual device name (Microsoft X-Box 360 pad), which cause this filter to not work. Also the device needs to be closed to actually stop recieving input events from it.